# openapi/opal-osa-integration.yaml
openapi: 3.0.0
info:
  title: OPAL Decision & Diagnostics API
  version: 1.0.0
  description: |
    Complete API specification for OPAL → OSA data flow, decision-making, and troubleshooting.

    This API maintains backward compatibility with existing OPAL integrations while providing
    enhanced decision layer capabilities with evidence-based recommendations, confidence scoring,
    and impact/effort analysis.

    **Key Features:**
    - Idempotent webhook processing for reliable data flow
    - Evidence-based decision making with confidence metrics
    - Comprehensive diagnostics and troubleshooting tools
    - Force sync and orchestration capabilities
    - Backward compatible with existing OPAL workflows
  contact:
    name: OPAL Development Team
    url: https://opal-2025.vercel.app
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://opal-2025.vercel.app
    description: Production server
paths:
  /api/orchestrations/trigger:
    post:
      tags:
        - Orchestration
      summary: Trigger a strategy workflow
      description: |
        Initiates a strategy workflow execution (Force Sync functionality).

        This endpoint maintains compatibility with existing OPAL orchestration
        flows while providing enhanced tracking and monitoring capabilities.
      operationId: triggerWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowTriggerRequest'
            examples:
              force_sync:
                summary: Force sync workflow trigger
                value:
                  strategy_assistant_workflow: "workflow-force-sync-2025"
                  trigger_source: "manual"
                  priority: "high"
              scheduled_sync:
                summary: Scheduled workflow trigger
                value:
                  strategy_assistant_workflow: "workflow-scheduled-sync-2025"
                  trigger_source: "scheduler"
                  priority: "normal"
      responses:
        '200':
          description: Workflow triggered successfully
          headers:
            X-Correlation-ID:
              schema:
                type: string
              description: Request correlation ID for tracing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowTriggerResponse'
              example:
                success: true
                data:
                  execution_id: "exec-1699876543210-abc123"
                  workflow_id: "workflow-force-sync-2025"
                  status: "initiated"
                  estimated_duration_minutes: 15
                correlation_id: "api-1699876543210-def456"
                timestamp: "2025-11-12T15:45:00.000Z"
                duration_ms: 89
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    WorkflowTriggerRequest:
      type: object
      required:
        - strategy_assistant_workflow
      properties:
        strategy_assistant_workflow:
          type: string
          description: Unique identifier for the workflow to trigger
          example: "workflow-force-sync-2025"
        trigger_source:
          type: string
          enum: [manual, scheduler, api, webhook]
          description: Source that initiated the workflow trigger
          default: "api"
        priority:
          type: string
          enum: [low, normal, high, critical]
          description: Execution priority level
          default: "normal"
        context:
          type: object
          description: Additional context data for workflow execution
          additionalProperties: true
...
    WorkflowTriggerResponse:
      allOf:
        - $ref: '#/components/schemas/StandardApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                execution_id:
                  type: string
                  description: Unique execution identifier
                  example: "exec-1699876543210-abc123"
                workflow_id:
                  type: string
                  description: Workflow identifier that was triggered
                status:
                  type: string
                  enum: [initiated, queued, running]
                  description: Current execution status
                estimated_duration_minutes:
                  type: integer
                  minimum: 1
                  description: Estimated completion time in minutes
...
    StandardApiResponse:
      type: object
      required:
        - success
        - correlation_id
        - timestamp
      properties:
        success:
          type: boolean
          description: Whether the request was successful
          example: true
        data:
          description: Response data (varies by endpoint)
        message:
          type: string
          description: Optional human-readable message
        correlation_id:
          type: string
          description: Unique request identifier for tracing
          example: "api-1699876543210-def456"
        timestamp:
          type: string
          format: date-time
          description: Response generation timestamp (ISO 8601 format)
          example: "2025-11-12T15:45:00.000Z"
        duration_ms:
          type: integer
          minimum: 0
          description: Processing time in milliseconds
          example: 245
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
        - correlation_id
        - timestamp
      properties:
        success:
          type: boolean
          description: Always false for error responses
          example: false
        error:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error description
          example: "Invalid request body"
        correlation_id:
          type: string
          description: Unique request identifier for tracing
          example: "api-1699876543210-def456"
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp (ISO 8601 format)
          example: "2025-11-12T15:45:00.000Z"
        duration_ms:
          type: integer
          minimum: 0
          description: Processing time before error occurred
          example: 25
        data:
          description: Additional error details (only in development)
  securitySchemes:
    hmacSignature:
      type: apiKey
      in: header
      name: X-Signature-SHA256
      description: |
        HMAC-SHA256 signature for webhook authentication.

        Format: sha256=<hex_digest>

        The signature is computed using the raw request body and a shared secret key.
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "VALIDATION_ERROR"
            message: "Invalid request body"
            correlation_id: "api-1699876543210-def456"
            timestamp: "2025-11-12T15:45:00.000Z"
            duration_ms: 25
    RateLimitExceeded:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "RATE_LIMIT_EXCEEDED"
            message: "Rate limit exceeded. Please retry after 60 seconds."
            correlation_id: "api-1699876543210-def456"
            timestamp: "2025-11-12T15:45:00.000Z"
            duration_ms: 5
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "INTERNAL_SERVER_ERROR"
            message: "An unexpected error occurred"
            correlation_id: "api-1699876543210-def456"
            timestamp: "2025-11-12T15:45:00.000Z"
            duration_ms: 150
...